requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
boto3==1.42.12
botocore==1.42.12
brotli==1.2.0
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2.9
django-allauth==65.13.1
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-filter==25.2
django-markdownx==4.0.9
django-multiselectfield==1.0.1
django-storages==1.14.6
django-stubs==5.2.8
django-stubs-ext==5.2.8
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
exceptiongroup==1.3.1
fonttools==4.61.1
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
Markdown==3.10
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psutil==7.1.3
pycparser==2.23
pydyf==0.12.1
PyJWT==2.10.1
pyphen==0.17.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
tinycss2==1.5.1
tinyhtml5==2.0.0
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
weasyprint==67.0
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.11.0
zopfli==0.4.0
